fix(struct): restructuration des fichiers pour récupérer les contribs
